+++ LauncherGo.Services/Extensions/ServiceCollectionExtension.cs
@@ -34,6 +34,7 @@ public static class ServiceCollectionExtension
         services.AddSingleton<IServerBridgeMigrationService, ServerBridgeMigrationService>();
         services.AddSingleton<Vs2QQProcessService>();
         services.AddSingleton<IRobotService, RobotService>();
+        services.AddSingleton<IDiscordBotService, DiscordBotService>();
         return services;
     }
 }